+++ src/main/java/zaftnotameni/creatania/mana/flowers/BotaniaFlowerInterfaces.java
@@ -1,0 +1,5 @@
+package zaftnotameni.creatania.mana.flowers;
+import vazkii.botania.api.block.IWandBindable;
+import vazkii.botania.api.block.IWandHUD;
+import vazkii.botania.api.block.IWandable;
+public interface BotaniaFlowerInterfaces extends IWandBindable, IWandable, IWandHUD { int doTick(); }